+++ package.json
@@ -37,6 +37,7 @@
 	"type": "module",
 	"dependencies": {
 		"@fontsource/ubuntu": "^5.0.8",
-		"fontsource-ubuntu": "^4.0.0"
+		"fontsource-ubuntu": "^4.0.0",
+		"yup": "^1.3.3"
 	}
 }
