+++ src/lib/Font.ts
@@ -1,0 +1,5 @@
+// place files you want to import through the `$lib` alias in this folder.
+import '@fontsource/ubuntu/300.css';
+import '@fontsource/ubuntu/400.css';
+import '@fontsource/ubuntu/500.css';
+import '@fontsource/ubuntu/700.css';
