+++ src/lib/stores/store-step.ts
@@ -1,6 +1,16 @@
 
 import { writable } from 'svelte/store';
+type Info = {
+	name: string;
+	email: string;
+	phone: string;
+};
 
+export const personInfo = writable<Info>({
+	name: '',
+	email: '',
+	phone: ''
+  });
 function createStepStore(){
 
 	const { subscribe, update} = writable(1);
